+++ src/main/java/aPresentation/Main.java
@@ -1,4 +1,4 @@
-package Presentation;
+package aPresentation;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
